refactor: Break down ThemeShadow into separate files
